Set Player stats Casting to uInt32

+++ src/game/Player.h
@@ -228,7 +228,7 @@ struct PlayerLevelInfo
 {
     PlayerLevelInfo() { for (int i = 0; i < MAX_STATS; ++i) stats[i] = 0; }
 
-    uint8 stats[MAX_STATS];
+    uint32 stats[MAX_STATS];
 };
 
 typedef std::list<uint32> PlayerCreateInfoSpells;
